Olympus lens transmittance
In tabular text format as Olympus provides only images on their website
not allowing a lot of calculations to be made with them

--

LUMPLFLN 40XW

--

The UMPLFLN, LUMPLFLN-W and LUMFLN series of long working distance,
water immersion objectives display flat images from high transmission
factors up to the near-infrared region of the spectrum. These objectives
achieve excellent DIC and fluorescence from the visible range to
infrared and are ideal for fluorescence imaging of brain tissue, as well
as other tissue and specimens.

--

XLPLN25XWMP2

--

These objectives are optimized for Multiphoton imaging and allow high
resolution 3D images to be obtained by detecting only the fluorescence
from the focal point of a big field of view.